refactor caper init (added caper_init.py)
