+++ caper/caper_args.py
@@ -11,13 +11,10 @@ import sys
 import os
 from distutils.util import strtobool
 from collections import OrderedDict
-from .caper_backend import BACKENDS, BACKENDS_WITH_ALIASES
-from .caper_backend import BACKEND_GCP, BACKEND_AWS, BACKEND_LOCAL
-from .caper_backend import BACKEND_SLURM, BACKEND_SGE, BACKEND_PBS
-from .caper_backend import BACKEND_ALIAS_LOCAL
-from .caper_backend import BACKEND_ALIAS_GOOGLE, BACKEND_ALIAS_AMAZON
-from .caper_backend import BACKEND_ALIAS_SHERLOCK, BACKEND_ALIAS_SCG
 from .caper_backend import CaperBackendDatabase
+from .caper_backend import BACKENDS, BACKEND_LOCAL
+from .caper_backend import BACKEND_ALIAS_LOCAL
+from .caper_backend import BACKEND_ALIAS_SHERLOCK, BACKEND_ALIAS_SCG
 
 
 __version__ = '0.6.5'
@@ -28,8 +25,6 @@ DEFAULT_CAPER_CONF = '~/.caper/default.conf'
 DEFAULT_SINGULARITY_CACHEDIR = '~/.caper/singularity_cachedir'
 DEFAULT_CROMWELL_JAR = 'https://github.com/broadinstitute/cromwell/releases/download/47/cromwell-47.jar'
 DEFAULT_WOMTOOL_JAR = 'https://github.com/broadinstitute/cromwell/releases/download/47/womtool-47.jar'
-DEFAULT_CROMWELL_JAR_INSTALL_DIR = '~/.caper/cromwell_jar'
-DEFAULT_WOMTOOL_JAR_INSTALL_DIR = '~/.caper/womtool_jar'
 DEFAULT_DB = CaperBackendDatabase.DB_TYPE_IN_MEMORY
 DEFAULT_MYSQL_DB_IP = 'localhost'
 DEFAULT_MYSQL_DB_PORT = 3306
@@ -54,67 +49,6 @@ DEFAULT_SERVER_HEARTBEAT_TIMEOUT_MS = 120000
 DEFAULT_CONF_CONTENTS = '\n\n'
 DYN_FLAGS = ['--singularity', '--docker']
 INVALID_EXT_FOR_DYN_FLAG = '.wdl'
-DEFAULT_CONF_CONTENTS_LOCAL = """backend=local
-
-# DO NOT use /tmp here
-# Caper stores all important temp files and cached big data files here
-tmp-dir=
-"""
-DEFAULT_CONF_CONTENTS_SHERLOCK = """backend=slurm
-slurm-partition=
-
-# DO NOT use /tmp here
-# Caper stores all important temp files and cached big data files here
-tmp-dir=
-"""
-DEFAULT_CONF_CONTENTS_SCG = """backend=slurm
-slurm-account=
-
-# DO NOT use /tmp here
-# Caper stores all important temp files and cached big data files here
-tmp-dir=
-"""
-DEFAULT_CONF_CONTENTS_SLURM = """backend=slurm
-
-# define one of the followings (or both) according to your
-# cluster's SLURM configuration.
-slurm-partition=
-slurm-account=
-
-# DO NOT use /tmp here
-# Caper stores all important temp files and cached big data files here
-tmp-dir=
-"""
-DEFAULT_CONF_CONTENTS_SGE = """backend=sge
-sge-pe=
-
-# DO NOT use /tmp here
-# Caper stores all important temp files and cached big data files here
-tmp-dir=
-"""
-DEFAULT_CONF_CONTENTS_PBS = """backend=pbs
-
-# DO NOT use /tmp here
-# Caper stores all important temp files and cached big data files here
-tmp-dir=
-"""
-DEFAULT_CONF_CONTENTS_AWS = """backend=aws
-aws-batch-arn=
-aws-region=
-out-s3-bucket=
-
-# DO NOT use /tmp here
-# Caper stores all important temp files and cached big data files here
-tmp-dir=
-"""
-DEFAULT_CONF_CONTENTS_GCP = """backend=gcp
-gcp-prj=
-out-gcs-bucket=
-
-# DO NOT use /tmp here
-# Caper stores all important temp files and cached big data files here
-tmp-dir=
-"""
 
 
 def process_dyn_flags(remaining_args, dyn_flags,
@@ -140,70 +74,6 @@ def process_dyn_flags(remaining_args, dyn_flags,
                     remaining_args[loc], remaining_args[loc + 1] = \
                         remaining_args[loc + 1], remaining_args[loc]
     return remaining_args
-
-
-def install_cromwell_jar(uri):
-    """Download cromwell-X.jar
-    """
-    from .caper_uri import CaperURI, URI_LOCAL
-    cu = CaperURI(uri)
-    if cu.uri_type == URI_LOCAL:
-        return cu.get_uri()
-    print('Downloading Cromwell JAR... {f}'.format(f=uri), file=sys.stderr)
-    path = os.path.join(
-        os.path.expanduser(DEFAULT_CROMWELL_JAR_INSTALL_DIR),
-        os.path.basename(uri))
-    return cu.copy(target_uri=path)
-
-
-def install_womtool_jar(uri):
-    """Download womtool-X.jar
-    """
-    from .caper_uri import CaperURI, URI_LOCAL
-    cu = CaperURI(uri)
-    if cu.uri_type == URI_LOCAL:
-        return cu.get_uri()
-    print('Downloading Womtool JAR... {f}'.format(f=uri), file=sys.stderr)
-    path = os.path.join(
-        os.path.expanduser(DEFAULT_WOMTOOL_JAR_INSTALL_DIR),
-        os.path.basename(uri))
-    return cu.copy(target_uri=path)
-
-
-def init_caper_conf(args):
-    """Initialize conf file for a given platform.
-    Also, download/install Cromwell/Womtool JARs.
-    """
-    backend = args.get('platform')
-    assert(backend in BACKENDS_WITH_ALIASES)
-    if backend in (BACKEND_LOCAL, BACKEND_ALIAS_LOCAL):
-        contents = DEFAULT_CONF_CONTENTS_LOCAL
-    elif backend == BACKEND_ALIAS_SHERLOCK:
-        contents = DEFAULT_CONF_CONTENTS_SHERLOCK
-    elif backend == BACKEND_ALIAS_SCG:
-        contents = DEFAULT_CONF_CONTENTS_SCG
-    elif backend == BACKEND_SLURM:
-        contents = DEFAULT_CONF_CONTENTS_SLURM
-    elif backend == BACKEND_SGE:
-        contents = DEFAULT_CONF_CONTENTS_SGE
-    elif backend == BACKEND_PBS:
-        contents = DEFAULT_CONF_CONTENTS_PBS
-    elif backend in (BACKEND_GCP, BACKEND_ALIAS_GOOGLE):
-        contents = DEFAULT_CONF_CONTENTS_GCP
-    elif backend in (BACKEND_AWS, BACKEND_ALIAS_AMAZON):
-        contents = DEFAULT_CONF_CONTENTS_AWS
-    else:
-        raise Exception('Unsupported platform/backend/alias.')
-
-    conf_file = os.path.expanduser(args.get('conf'))
-    with open(conf_file, 'w') as fp:
-        fp.write(contents + '\n')
-        fp.write('{key}={val}\n'.format(
-            key='cromwell',
-            val=install_cromwell_jar(DEFAULT_CROMWELL_JAR)))
-        fp.write('{key}={val}\n'.format(
-            key='womtool',
-            val=install_womtool_jar(DEFAULT_WOMTOOL_JAR)))
 
 
 def parse_caper_arguments():
@@ -674,11 +544,5 @@ def parse_caper_arguments():
         v = args_d.get(k)
         if v is not None and isinstance(v, str):
             args_d[k] = int(v)
-
-    # if action is 'init' then initialize Conf and exit
-    action = args_d['action']
-    if action == 'init':
-        init_caper_conf(args_d)
-        sys.exit(0)
 
     return args_d
